requirements.txt
asgiref==3.5.2
Django==4.0.5
django-versatileimagefield==2.2
Pillow==9.1.1
python-magic==0.4.27
sqlparse==0.4.2
tzdata==2022.1
